Update pip install command
The old command failed to install the dependency. Running the command returned error:
```
Downloading/unpacking version==1.12.0-2
Could not find a version that satisfies the requirement version==1.12.0-2 (from versions: 0.1.0, 0.1.1)
Cleaning up...
No distributions matching the version for version==1.12.0-2
```
When asked about the issue via IRC @atnnn told me that the error is not OS dependent, so therefore I deiced to file a pull request for it.