Fixed inconsistency in table_create api docs
All three versions of the API docs state that trying to create a table
with a name that duplicates an existing table will raise a
ReqlRuntimeError. This contradicts the test at lines 132-136 of
test/rql_test/src/meta/table.yaml which specifies that in this case a
ReqlOpFailedError should be raised.

Docs have been updated to match behavior specified by test.