Problem: current `Response` API isn't as nice as it could be
The current `Response` is just a typedef for
`futures::sync::mpsc::Receiver`. While the futures channels are
very useful for sending information between threads, returning
the `Receiver` directly is less than ideal. This is because
`Stream::Error` is set to `()` which means that `Stream::Item`
has to be a `Result`. This effectively makes any Stream methods
that work on errors pretty much useless.

Solution: hide the `Receiver` and `impl Stream for Response`.