Problem: token not in stream cache
This is an error message which comes from RethinkDB. This was caused by
a logic bug where we encoded a the query before retrying connections.
This meant that RethinkDB ended up getting a new token ID in a request
to retrieve data yet an old ID was used when submitting the message.

Solution: make sure to encode data _after_ retrying a connection