Allow creating index using binary representation
While building an alternative import tool for rethinkdb using Rust, I
found myself trying to create index using the binary representation, as
exported by the official export tool.

However, the current implementation always wrap everything in a `Func`
so that didn't work, this PR is an attempt to fix that.

There's probably a better way to do this, but I'm not sure how.